requirements.txt
Flask
flask-jwt-extended
flask-bcrypt
peewee
PyMySQL
yoyo-migrations
boto3